@@ -2,14 +2,15 @@
   "name": "shikimori",
   "private": true,
   "dependencies": {
-    "@babel/plugin-proposal-decorators": "^7.13.15",
-    "@babel/plugin-proposal-logical-assignment-operators": "^7.13.8",
-    "@babel/plugin-proposal-optional-chaining": "^7.12.7",
+    "@babel/plugin-proposal-decorators": "^7.14.2",
+    "@babel/plugin-proposal-logical-assignment-operators": "^7.14.2",
+    "@babel/plugin-proposal-optional-chaining": "^7.14.2",
     "@babel/plugin-proposal-partial-application": "^7.12.13",
     "@babel/plugin-proposal-pipeline-operator": "^7.12.1",
-    "@johmun/vue-tags-input": "^2.1.0",
     "@morr/user-idle": "^2.0.0",
-    "@rails/webpacker": "5.2.1",
+    "@rails/webpacker": "5.4.0",
+    "@sipec/vue3-tags-input": "3.0.3",
+    "@vue/compiler-sfc": "^3.0.11",
     "autosize": "^4.0.2",
     "axios": "^0.21.1",
     "blueimp-md5": "2.18.0",
@@ -18,6 +19,7 @@
     "dayjs": "^1.10.4",
     "delay": "^5.0.0",
     "diff-match-patch": "^1.0.5",
+    "fast-async": "7",
     "faye": "^1.3.0",
     "get-js": "^0.0.7",
     "highcharts": "^9.0.0",
@@ -42,7 +44,7 @@
     "pug-loader": "^2.4.0",
     "pug-plain-loader": "^1.1.0",
     "shiki-decorators": "^1.0.1",
-    "shiki-editor": "^1.3.3",
+    "shiki-editor": "^2.0.4",
     "shiki-uploader": "^1.1.1",
     "shiki-utils": "^1.0.1",
     "swiper": "^6.5.8",
@@ -51,21 +53,20 @@
     "turbolinks": "^5.2.0",
     "uevent": "^2.0.1",
     "url-parse": "^1.5.1",
-    "vue": "2.6.12",
-    "vue-loader": "^15.9.3",
-    "vue-template-compiler": "2.6.12",
-    "vuedraggable": "^2.24.0",
-    "vuex": "^3.6.2",
+    "vue": "^3.0.11",
+    "vue-loader": "^16.2.0",
+    "vuedraggable": "^4.0.1",
+    "vuex": "^4.0.0",
     "webfontloader": "^1.6.28"
   },
   "devDependencies": {
     "babel-eslint": "11.0.0-beta.2",
     "es-check": "^5.2.3",
-    "eslint": "^7.25.0",
+    "eslint": "^7.27.0",
     "eslint-config-airbnb-base": "^14.2.1",
-    "eslint-import-resolver-webpack": "0.13.0",
-    "eslint-plugin-import": "^2.22.1",
-    "eslint-plugin-vue": "6.2.2",
+    "eslint-import-resolver-webpack": "^0.13.1",
+    "eslint-plugin-import": "^2.23.3",
+    "eslint-plugin-vue": "7.9.0",
     "fsevents": "^2.3.2",
     "webpack-bundle-analyzer": "^4.4.1",
     "webpack-dev-server": "^3.11.0",
