@@ -35,29 +35,24 @@ pageLoad(
   }
 );
 
-// sort with preserving initial order
-// function sortByGroups(data) {
-//   data.links = [].concat.apply([], Object.values(data.links.groupBy(v => v.group)));
-//   return data;
-// }
-
 async function initVueApp() {
   if (!$('#vue_collection_links').exists()) { return; }
 
-  const { Vue, Vuex } = await import(/* webpackChunkName: "vue" */ '@/vue/instance');
+  const { createApp } = await import(/* webpackChunkName: "vue" */ 'vue');
+  const { createStore } = await import(/* webpackChunkName: "vuex" */ 'vuex');
   const { default: CollectionLinks } = await import('@/vue/components/collections/collection_links'); // eslint-disable-line max-len
+
   const storeSchema = await import('@/vue/stores/collection_links');
 
   const collection = $('#collection_form').data('collection');
   const autocompleteUrl = $('#collection_form').data('autocomplete_url');
   const maxLinks = $('#collection_form').data('max_links');
 
-  const store = new Vuex.Store(storeSchema);
-  store.state.collection = collection; // sortByGroups(collection)
+  const store = createStore(storeSchema);
+  store.state.collection = collection;
 
-  new Vue({
-    el: '#vue_collection_links',
-    store,
-    render: h => h(CollectionLinks, { props: { maxLinks, autocompleteUrl } })
-  });
+  const app = createApp(CollectionLinks, { maxLinks, autocompleteUrl });
+  app.use(store);
+  app.config.globalProperties.I18n = I18n;
+  app.mount('#vue_collection_links');
 }
