@@ -1,13 +1,15 @@
 pageLoad('.polls', async () => {
   if (!$('#vue_poll_variants').exists()) { return; }
 
-  const { Vue, Vuex } = await import(/* webpackChunkName: "vue" */ '@/vue/instance');
+  const { createApp } = await import(/* webpackChunkName: "vue" */ 'vue');
+  const { createStore } = await import(/* webpackChunkName: "vuex" */ 'vuex');
+
   const { default: ArrayField } = await import('@/vue/components/array_field');
   const { default: storeSchema } = await import('@/vue/stores/collection');
 
   const pollVariants = $('#poll_form').data('poll').variants;
 
-  const store = new Vuex.Store(storeSchema);
+  const store = createStore(storeSchema);
   store.state.collection = pollVariants.map((pollVariant, index) =>
     ({
       key: index,
@@ -15,16 +17,13 @@ pageLoad('.polls', async () => {
     })
   );
 
-  new Vue({
-    el: '#vue_poll_variants',
-    store,
-    render: h => h(ArrayField, {
-      props: {
-        resourceType: 'Poll',
-        field: 'poll_variants',
-        inputName: 'poll[variants_attributes][][label]',
-        emptyInputName: 'poll[variants_attributes][]'
-      }
-    })
+  const app = createApp(ArrayField, {
+    resourceType: 'Poll',
+    field: 'poll_variants',
+    inputName: 'poll[variants_attributes][][label]',
+    emptyInputName: 'poll[variants_attributes][]'
   });
+  app.use(store);
+  app.config.globalProperties.I18n = I18n;
+  app.mount('#vue_poll_variants');
 });
