@@ -1,99 +1,99 @@
 <template lang='pug'>
-  .b-input
-    label
-      | {{ label }}
-      vue-tags-input(
-        :add-on-key='addOnKey'
-        :separators='separators'
-        :autocomplete-items='autocompleteItems'
-        :autocomplete-always-open='tags.length < tagsLimit && !!autocompleteItems.length'
-        :tags='tags'
-        :max-tags='tagsLimit'
-        @before-adding-tag='checkTag'
-        @tags-changed='syncToInput'
-        placeholder=''
-        v-model='tag'
-      )
-    span.hint(
-      v-if='hint'
-      v-html='hint'
+.b-input
+  label
+    | {{ label }}
+    VueTagsInput(
+      :add-on-key='addOnKey'
+      :separators='separators'
+      :autocomplete-items='autocompleteItems'
+      :autocomplete-always-open='tags.length < tagsLimit && !!autocompleteItems.length'
+      :tags='tags'
+      :max-tags='tagsLimit'
+      @before-adding-tag='checkTag'
+      @tags-changed='syncToInput'
+      placeholder=''
+      v-model='tagModel'
     )
+  span.hint(
+    v-if='hint'
+    v-html='hint'
+  )
 </template>
 
-<script>
-import VueTagsInput from '@johmun/vue-tags-input';
+<script setup>
+import { defineProps, ref, computed } from 'vue';
 
-export default {
-  name: 'TagsInput',
-  components: { VueTagsInput },
-  props: {
-    label: { type: String, required: true },
-    hint: { type: String, required: false, default: undefined },
-    tagsLimit: { type: Number, required: true },
-    autocompleteBasic: { type: Array, required: true },
-    autocompleteOther: { type: Array, required: true },
-    input: { type: HTMLInputElement, required: true },
-    value: { type: Array, required: true },
-    isDowncase: { type: Boolean, required: false, default: false }
-  },
-  data() {
-    return {
-      tag: '',
-      tags: this.value.map(v => ({ text: v }))
-    };
-  },
-  computed: {
-    autocompleteItems() {
-      return (
-        this.tags.length && this.autocompleteOther.length ?
-          this.autocompleteOther :
-          this.autocompleteBasic
-      )
-        .filter(v => !this.tags.find(tag => tag.text === v))
-        .filter(v => (this.tag ? v.startsWith(this.tag) : true))
-        .map(v => ({ text: v }));
-    },
-    separators() {
-      return this.tagsLimit > 1 ? [';', ',', ' '] : undefined;
-    },
-    addOnKey() {
-      return this.tagsLimit > 1 ? [9, 13, 32, ','] : undefined;
-    }
-  },
-  methods: {
-    checkTag({ tag, addTag }) {
-      if (this.isDowncase) {
-        tag.text = tag.text.toLowerCase();
-      }
-      addTag(tag);
-    },
-    syncToInput(newTags) {
-      this.tags = newTags;
-      this.input.value = this.tags.map(v => v.text).join(',');
-    }
+import VueTagsInput from '@sipec/vue3-tags-input';
+
+const props = defineProps({
+  label: { type: String, required: true },
+  hint: { type: String, required: false, default: undefined },
+  tagsLimit: { type: Number, required: true },
+  autocompleteBasic: { type: Array, required: true },
+  autocompleteOther: { type: Array, required: true },
+  input: { type: HTMLInputElement, required: true },
+  value: { type: Array, required: true },
+  isDowncase: { type: Boolean, required: false, default: false }
+});
+
+let tagModel = ref('');
+let tags = ref(props.value.map(v => ({ text: v })));
+
+const autocompleteItems = computed(() => (
+  (
+    tags.value.length && props.autocompleteOther.length ?
+      props.autocompleteOther :
+      props.autocompleteBasic
+  )
+    .filter(v => !tags.value.find(tag_value => tag_value.text === v))
+    .filter(v => (tagModel.value ? v.startsWith(tagModel.value) : true))
+    .map(v => ({ text: v }))
+));
+
+const separators = computed(() => (
+  props.tagsLimit > 1 ?
+    [';', ',', ' '] :
+    undefined
+));
+
+const addOnKey = computed(() => (
+  props.tagsLimit > 1 ?
+    [9, 13, 32, ','] :
+    undefined
+));
+
+function checkTag({ tag, addTag }) {
+  if (props.isDowncase) {
+    tag.text = tag.text.toLowerCase();
   }
-};
+  addTag(tag);
+}
+
+function syncToInput(newTags) {
+  tags.value = newTags;
+  props.input.value = tags.value.map(v => v.text).join(','); // eslint-disable-line vue/no-mutating-props
+}
 </script>
 
 <style scoped lang='sass'>
 @import 'app/assets/stylesheets/mixins/input'
 
-.b-input /deep/
-  .vue-tags-input
+.b-input
+  ::v-deep(.vue-tags-input)
     max-width: 100% !important
 
-  .ti-autocomplete
+  ::v-deep(.ti-autocomplete)
     display: none
     z-index: 31
 
-  .ti-focus
+  ::v-deep(.ti-focus)
     .ti-input
       +input_focus
 
     .ti-autocomplete
       display: block
 
-  .ti-input
+  ::v-deep(.ti-input)
     +input
     padding: 0 1px
     max-width: 100% !important
@@ -101,7 +101,7 @@ export default {
     input
       width: 100% !important
 
-  .ti-tag
+  ::v-deep(.ti-tag)
     font-size: 12px
     margin: 4px 2px
 
@@ -119,7 +119,7 @@ export default {
     .ti-content
       align-items: flex-start
 
-  .ti-new-tag-input-wrapper
+  ::v-deep(.ti-new-tag-input-wrapper)
     margin: 0
     padding: 0
 
