@@ -1,59 +1,59 @@
 <template lang='pug'>
-  .b-collection_item(
-    :data-linked_id='link.linked_id'
-    :data-group='link.group'
-    :data-list_index='links.indexOf(link)'
-  )
-    div
-      .delete(
-        @click="removeLink(link)"
-      )
-      .drag-handle(
-        v-if='link.linked_id'
-      )
-    .b-input.new-record(
-      v-if='!link.linked_id'
+.b-collection_item(
+  :data-linked_id='link.linked_id'
+  :data-group='link.group'
+  :data-list_index='links.indexOf(link)'
+)
+  div
+    .delete(
+      @click="removeLink(link)"
     )
-      label
-        //| {{ I18n.t(`frontend.collections.add.${collection.kind}`) }}
-        input(
-          type='text'
-          :placeholder="I18n.t(`frontend.collections.autocomplete.${collection.kind}`)"
-          :data-autocomplete='autocompleteUrl'
-        )
-    .persisted(
+    .drag-handle(
       v-if='link.linked_id'
     )
+  .b-input.new-record(
+    v-if='!link.linked_id'
+  )
+    label
+      //| {{ I18n.t(`frontend.collections.add.${collection.kind}`) }}
       input(
-        v-if='link.id'
-        type='hidden'
-        v-model='link.id'
-        name='collection[links][][id]'
+        type='text'
+        :placeholder="I18n.t(`frontend.collections.autocomplete.${collection.kind}`)"
+        :data-autocomplete='autocompleteUrl'
       )
-      input(
-        type='hidden'
-        v-model='link.linked_id'
-        v-if='link.linked_id'
-        name='collection[links][][linked_id]'
+  .persisted(
+    v-if='link.linked_id'
+  )
+    input(
+      v-if='link.id'
+      type='hidden'
+      v-model='link.id'
+      name='collection[links][][id]'
+    )
+    input(
+      type='hidden'
+      v-model='link.linked_id'
+      v-if='link.linked_id'
+      name='collection[links][][linked_id]'
+    )
+    input(
+      type='hidden'
+      v-model='link.group'
+      name='collection[links][][group]'
+    )
+    a.b-link(
+      :class="link.url ? 'bubbled' : ''"
+      :href='link.url'
+      data-predelay='600'
+    ) {{ link.name }}
+    .b-input
+      textarea(
+        :placeholder="I18n.t('activerecord.attributes.collection_link.text')"
+        name='collection[links][][text]'
+        rows='1'
+        v-model='link.text'
+        @focus.once='add_autosize'
       )
-      input(
-        type='hidden'
-        v-model='link.group'
-        name='collection[links][][group]'
-      )
-      a.b-link(
-        :class="link.url ? 'bubbled' : ''"
-        :href='link.url'
-        data-predelay='600'
-      ) {{ link.name }}
-      .b-input
-        textarea(
-          :placeholder="I18n.t('activerecord.attributes.collection_link.text')"
-          name='collection[links][][text]'
-          rows='1'
-          v-model='link.text'
-          @focus.once='add_autosize'
-        )
 </template>
 
 <script>
