@@ -1,34 +1,34 @@
 <template lang='pug'>
-  .block
-    input(
-      type='hidden'
-      :name="`${resourceType.toLowerCase()}[external_links][]`"
-      v-if='isEmpty'
-    )
-    .b-nothing_here(
-      v-if='!collection.length'
-    )
-      | {{ I18n.t('frontend.external_links.nothing_here') }}
-    draggable.block(
-      v-bind='dragOptions'
-      v-model='collection'
-      v-if='collection.length'
-    )
+.block_m
+  input(
+    type='hidden'
+    :name="`${resourceType.toLowerCase()}[external_links][]`"
+    v-if='isEmpty'
+  )
+  .b-nothing_here(
+    v-if='!collection.length'
+  )
+    | {{ I18n.t('frontend.external_links.nothing_here') }}
+  draggable.block(
+    v-if='collection.length'
+    v-model='collection'
+    item-key='element => element.id || element.key'
+    v-bind='dragOptions'
+  )
+    template(#item="{element}")
       ExternalLink(
-        v-for='link in collection'
-        @add_next='add'
-        @focusLast='focusLast'
-        :key='link.id || link.key'
-        :link='link'
+        :link='element'
         :kind-options='kindOptions'
         :resource-type='resourceType'
         :entry-type='entryType'
         :entry-id='entryId'
         :watch-online-kinds='watchOnlineKinds'
+        @add:next='add'
+        @focusLast='focusLast'
       )
-    .b-button(
-      @click='add'
-    ) {{ I18n.t('frontend.actions.add') }}
+  .b-button(
+    @click='add'
+  ) {{ I18n.t('frontend.actions.add') }}
 </template>
 
 <script>
@@ -87,6 +87,6 @@ export default {
 </script>
 
 <style scoped lang='sass'>
-  .b-nothing_here
-    margin-bottom: 15px
+.b-nothing_here
+  margin-bottom: 15px
 </style>
