@@ -1,6 +1,6 @@
 if defined? Bugsnag
   Bugsnag.configure do |config|
-    config.api_key = 'ddc5a1c19acba6ee8c734febc5af5dbf'
+    config.api_key = '91511aee24995957c05b0f51e187dd39'
 
     Shikimori::IGNORED_EXCEPTIONS
       .map { |v| v.constantize rescue NameError }
